feat: Add support for background processes (&)
This commit introduces the ability to run commands in the background by appending &. The shell will no longer wait for backgrounded commands and will print the PID of the new process. This also includes a zombie reaping mechanism to clean up completed background jobs. The existing, partially-working pipe implementation is included in this commit.